Rename AI hooks
